Move PRISM to Git submodule
Rather than relying on OS/architecture-specific libraries in
lib/prism-*, include PRISM as a submodule of the repository. This makes
it much less of a hassle to compile SCH-IMP on non-linux64 platforms.

ext/prism/ tracks a specific branch of my fork of the prism-ext
repository, so Git >= 1.8.2 is required for a full clone from now on.