Dynamically set Java library path when launched from JAR file
PRISM depends on native-code libraries that need to be available when
the PRISM engine starts; this means that the directory containing these
libraries needs to be present in the Java library path
(java.library.path) when PRISM attempts to load them. Traditionally the
user would be expected to set java.library.path manually on the command
line, but if SCH-IMP is being run via its JAR file, assume that the
native-code libraries are in the same directory as the JAR file (per the
"dist" Ant target) and automatically prepend this directory to
java.library.path during initialisation.

Note that java.library.path must still be set if running CommandLine
from outside of a JAR file built by "ant dist", and the system's loader
must still be able to find libraries in non-standard paths that the
native-code libraries themselves depend on (e.g. LD_LIBRARY_PATH will
still need to be set in Linux).